refactor: reorganize UI constants from category-based to feature-based modules
Previously, constants were split by category (column_headers, table_dimensions,
table_styles), requiring each renderer/widget to import from 3 separate files.
Now constants are grouped by feature (gantt, task_table, audit_log, timeline)
with shared cross-cutting constants in common.py.

- Add common.py for shared constants (table styles, ID/Name/Estimated headers)
- Add gantt.py, task_table.py, audit_log.py, timeline.py for feature-specific constants
- Add formatting.py for format_table_title() utility
- Move BACKGROUND_COLOR_* from symbols.py to colors.py (proper responsibility)
- Move gantt-specific colors from colors.py to gantt.py
- Extract timeline constants from local definitions in presenters/renderers
- Delete column_headers.py, table_dimensions.py, table_styles.py
- Update all 17 consumer files with new import paths
