--- packages/taskdog-ui/src/taskdog/constants/column_headers.py
@@ -1,38 +1,0 @@
-"""Column header constants shared between CLI and TUI."""
-
-# Task Table headers
-HEADER_ID = "ID"
-HEADER_NAME = "Name"
-HEADER_STATUS = "Status"
-HEADER_PRIORITY = "Priority"
-HEADER_FLAGS = "Flags"
-HEADER_ESTIMATED = "Estimated[h]"
-HEADER_ACTUAL = "Actual[h]"
-HEADER_DEADLINE = "Deadline"
-HEADER_PLANNED_START = "Planned Start"
-HEADER_PLANNED_END = "Planned End"
-HEADER_ACTUAL_START = "Actual Start"
-HEADER_ACTUAL_END = "Actual End"
-HEADER_ELAPSED = "Elapsed"
-HEADER_DEPENDENCIES = "Dependencies"
-HEADER_TAGS = "Tags"
-HEADER_NOTE = "Note"
-HEADER_FIXED = "Fixed"
-HEADER_DURATION = "Duration"
-HEADER_CREATED_AT = "Created At"
-HEADER_UPDATED_AT = "Updated At"
-
-# Gantt-specific headers
-HEADER_TIMELINE = "Timeline"
-
-# Gantt workload row label
-GANTT_WORKLOAD_LABEL = "Est. Workload[h]"
-
-# Audit Log Table headers
-HEADER_AUDIT_TIMESTAMP = "Timestamp"
-HEADER_AUDIT_CLIENT = "Client"
-HEADER_AUDIT_OPERATION = "Operation"
-HEADER_AUDIT_RESOURCE = "Resource"
-HEADER_AUDIT_STATUS = "Status"
-HEADER_AUDIT_CHANGES = "Changes"
-HEADER_AUDIT_STATUS_SHORT = "St"
