+++ packages/taskdog-ui/src/taskdog/constants/symbols.py
@@ -1,4 +1,4 @@
-"""Symbol and background color constants for Gantt chart rendering."""
+"""Symbol constants for Gantt chart rendering."""
 
 # Visual constants for Gantt chart symbols
 SYMBOL_PLANNED = "░"
@@ -16,11 +16,3 @@ SYMBOL_CANCELED = "x"  # lowercase x - task canceled
 
 # UI Emojis
 EMOJI_NOTE = "📝"  # Note indicator in task table
-
-# Background colors for Gantt chart
-BACKGROUND_COLOR = "rgb(100,100,100)"  # Weekday (allocated hours)
-BACKGROUND_COLOR_SATURDAY = "rgb(100,100,150)"  # Saturday (blueish, allocated hours)
-BACKGROUND_COLOR_SUNDAY = "rgb(150,100,100)"  # Sunday (reddish, allocated hours)
-BACKGROUND_COLOR_HOLIDAY = "rgb(200,150,100)"  # Holiday (orange-ish, allocated hours)
-BACKGROUND_COLOR_PLANNED_LIGHT = "rgb(60,60,60)"  # Planned period (no allocation yet)
-BACKGROUND_COLOR_DEADLINE = "rgb(200,100,0)"  # Deadline (orange)
