--- packages/taskdog-ui/src/taskdog/constants/table_dimensions.py
@@ -1,38 +1,0 @@
-"""Table and UI dimension constants for presentation layer."""
-
-# Task Table Display Limits
-TASK_NAME_COLUMN_WIDTH = 30
-STATUS_COLUMN_WIDTH = 13
-ESTIMATED_COLUMN_WIDTH = 14
-PAGE_SCROLL_SIZE = 10
-
-# Gantt Chart Column Widths (CLI)
-GANTT_TABLE_ID_WIDTH = 4
-GANTT_TABLE_TASK_MIN_WIDTH = 20
-GANTT_TABLE_EST_HOURS_WIDTH = 7
-
-# Gantt Chart Timeline Dimensions
-MIN_TIMELINE_WIDTH = 30
-CHARS_PER_DAY = 3
-# Fixed columns overhead for CLI Gantt: ID(4) + Name(30) + Est(7) + padding(8) + borders(5)
-GANTT_CLI_FIXED_COLUMNS_WIDTH = 54
-
-# Gantt Widget Dimensions (TUI)
-DEFAULT_GANTT_WIDGET_WIDTH = 120
-MIN_CONSOLE_WIDTH = 80
-BORDER_WIDTH = 2
-MIN_GANTT_DISPLAY_DAYS = 91
-
-# Audit Log Table Dimensions (CLI)
-AUDIT_ID_WIDTH = 6
-AUDIT_TIMESTAMP_WIDTH = 19
-AUDIT_CLIENT_WIDTH = 15
-AUDIT_OPERATION_WIDTH = 18
-AUDIT_CHANGES_WIDTH = 30
-AUDIT_STATUS_WIDTH = 8
-
-# Audit Log Table Dimensions (TUI)
-AUDIT_TUI_ID_WIDTH = 5
-AUDIT_TUI_NAME_WIDTH = 25
-AUDIT_TUI_CHANGES_WIDTH = 40
-AUDIT_TUI_STATUS_WIDTH = 4
